backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pillow==10.2.0
numpy==1.26.4
